+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Problems with loading physical values
 - Incorrect warning generation
 - Incorrect mood episode detection
+- Old list of scale value categories
 - Cleaned up code and tests
 
 ## [1.0.0-beta.1] - 2026-08-15